Generalized plugins: removed triggers, added handlers for non-message events
More specifically, the single run() method for plugins is now obsolete, as are "triggers".
Plugins define methods for whatever they're intended to handle now (msg, join, part, and so on)
and all events of that type are passed to the method. This implies that plugin methods are now
responsible for filtering messages themselves, since that logic has been removed from Cassium.